+++ Frontend/src/page/UserManagement.jsx
@@ -1,193 +1,15 @@
-import { useState, useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useEffect } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { getAllUsers } from "../features/auth/adminDashboardSlice";
 
-import {
-    Activity,
-    ArrowUpRight,
-    BookOpen,
-    CircleCheckBig,
-    Clock3,
-    FileText,
-    GraduationCap,
-    Layers3,
-    ShieldCheck,
-    Sparkles,
-    Users,
-    UserStar,
-    UserSearch,
-    Trash2,
-} from 'lucide-react'
-import { getAllUsers } from '../features/auth/adminDashboardSlice';
-
-function UsersHeader() {
-    
-    const stats = [
-        {
-            title: "Total Users",
-            value: "3",
-            icon: Users,
-            tone: "from-cyan-500 to-blue-500",
-        },
-
-        {
-            title: "Admin Users",
-            value: "2",
-            icon: UserStar,
-            tone: "from-emerald-500 to-teal-500",
-        }
-    ]
-    
-    return(
-        <>
-            <div className='mx-auto flex w-full max-w-7xl flex-col gap-6 p-4 pt-20 md:p-6 md:pt-6 lg:p-8'>
-                <section className='relative overflow-hidden rounded-[2rem] border border-slate-200 bg-white p-6 shadow-[0_24px_80px_rgba(15,23,42,0.08)] md:p-8'>
-                    <div className="absolute inset-0 bg-[radial-gradient(circle_at_top_right,_rgba(34,211,238,0.16),_transparent_30%),radial-gradient(circle_at_bottom_left,_rgba(59,130,246,0.10),_transparent_28%)]" />
-                    
-                    <div className='relative flex flex-col gap-6 lg:flex-row lg:items-center lg:justify-between'>
-                        <div className='max-w-2xl space-y-4'>
-                            <div className="inline-flex items-center gap-2 rounded-full border border-cyan-100 bg-cyan-50 px-4 py-2 text-sm text-cyan-700">
-                                <Sparkles size={16} />
-                                All User Managment Page
-                            </div>
-
-                            <div className='space-y-5'>
-                                <h1 className='roboto-bold text-2xl leading-tight text-slate-900 md:text-3xl'>User Managment Managme Stap-Wise Users</h1>
-                            </div>
-                        </div>
-                    </div>
-                </section>
-
-                <section className='grid gap-4 md:grid-cols-2'>
-                    {stats.map((stat) => {
-                        const Icon = stat.icon;
-
-                        return (
-                            <div key={stat.title} className='relative rounded-3xl border border-slate-200 bg-white p-5 shadow-sm transition-transform duration-300 hover:-translate-y-1'>
-                                <div className='flex items-start justify-between gap-4'>
-                                    <div>
-                                        <p className='text-sm text-slate-500'>{stat.title}</p>
-                                        <h2 className='mt-3 roboto-bold text-3xl text-slate-900'>{stat.value}</h2>
-                                    </div>
-
-                                    <div className={`flex h-12 w-12 items-center justify-center rounded-2xl bg-gradient-to-br ${stat.tone} text-white shadow-lg`}>
-                                        <Icon size={22} />
-                                    </div>
-                                </div>
-                            </div>
-                        )
-                    })}
-                </section>
-            </div>
-        </>
-    ) 
-}
-
-function UsersTable({users, isLoading, Loading}) {
-
-    // const fullName = user?.fullName || "Admin";
-    // const initials = fullName
-    //     .split(" ")
-    //     .filter(Boolean)
-    //     .slice(0, 2)
-    //     .map((name) => name[0]?.toUpperCase())
-    //     .join("") || "AD";
-
-    const TableRows = [
-        {
-            fullName: "Full Name",
-            email: "Email",
-            skill: "Skill",
-            enroled: "En-roled",
-            role: "Role"
-        },
-    ]
-
-   useEffect(() => {
-        if (users) {
-            users.forEach((user) => {
-                console.log(user.full_name);
-            });
-        }
-    }, [users]);
-
-    const userItems = [
-        {
-            name: isLoading ? "Loading..." : (users?.full_name),
-            email: "neftalemdagnachew@gmail.com",
-            skill: "Web-devlopment",
-            enroled: "True",
-            role: "User",
-            icon: Trash2,
-            delete: "Delete User",
-        }
-    ]
-
-    return(
-        <>
-            <div className='mx-auto flex w-full max-w-7xl flex-col gap-6 p-4 pt-20 md:p-6 md:pt-6 lg:p-8'>
-                <div className='w-full max-w-7xl bg-white relative rounded-3xl p-8 md:p-10 lg:p-12'>
-                    <div className='absolute rounded-3xl inset-0 bg-[radial-gradient(circle_at_bottom_right,_rgba(34,211,238,0.16),_transparent_20%),radial-gradient(circle_at_top_left,_rgba(59,130,246,0.10),_transparent_18%)]' />
-
-                    <div className='flex items-center justify-end'>
-                        <div className='relative flex items-center'>
-                            <input type="text" placeholder='Search' className='p-2 md:px-5 border border-slate-400 rounded-xl shadow-md outline-hidden' />
-                            <label className='absolute right-3 cursor-pointer'><UserSearch size={20} /></label>
-                        </div>
-                    </div>
-                    <div className='mt-5 md:mt-3'>
-                        <h4 className='roboto-bold text-xl leading-tight text-slate-900 md:text-3xl'>All Stap-Wise Users</h4>
-                    </div>
-                    <div className='relative mt-5 w-full'>
-                        <div className='overflow-x-auto'>
-                            <div className='p-4 md:p-6 min-w-[1120px] border border-slate-200 rounded-2xl shadow-[0_24px_80px_rgba(15,23,42,0.08)]'>
-                                
-                                <header className='w-full bg-slate-50 p-5 rounded-2xl'>
-                                    {TableRows.map((TableRow, index) => {
-                                        return(
-                                            <div key={index} className='grid grid-cols-7 gap-4'>
-                                                <p className='roboto-medium col-span-2 text-lg'>{TableRow.fullName}</p>
-                                                <p className='roboto-medium col-span-2 text-lg'>{TableRow.email}</p>
-                                                <p className='roboto-medium text-lg'>{TableRow.skill}</p>
-                                                <p className='roboto-medium text-lg'>{TableRow.enroled}</p>
-                                                <p className='roboto-medium text-lg'>{TableRow.role}</p>
-                                            </div>
-                                        )
-                                    })}
-                                </header>
-
-                                <div className='relative p-4 md:p-6 grid grid-cols-7 items-center gap-4'>
-                                    <div className='col-span-2 flex items-center gap-1'>
-                                        <div className="flex h-11 w-11 shrink-0 items-center justify-center rounded-full bg-gradient-to-br from-cyan-400 to-blue-500 text-sm roboto-bold text-white">
-                                            ND
-                                        </div>
-                                        <p className=' roboto-semibold text-xl'>Neftalem Dagnachew</p>
-                                    </div>
-                                    <p className='col-span-2 roboto-regular'>@neftalemdagnahcew</p>
-                                    <p className='roboto-regular'>Web-devlopment</p>
-                                    <p className='roboto-regular'>True</p>
-                                    <p className='roboto-regular'>User</p>
-                                    <div className='flex absolute right-1 items-center text-rose-600 border border-transparent hover:border-rose-200 hover:bg-rose-50 px-3 py-2 cursor-pointer rounded-lg transition-all duration-300'>
-                                        <span className='pr-1'> <Trash2 size={20} /> </span>
-                                        <button className='cursor-pointer'>Delete</button>
-                                    </div>
-                                </div>
-
-                            </div>
-
-                        </div>
-                    </div>
-
-                </div>
-            </div>
-        </>
-    )
-
-}
+// components
+import UsersHeader from "../components/users/UsersHeader";
+import UsersTable from "../components/users/UsersTable";
 
 function UserManagement() {
     const dispatch = useDispatch();
-    
-    const { users, isLoading, isError, message } = useSelector((state) => state.admin);
+
+    const { users, isLoading } = useSelector((state) => state.admin);
 
     useEffect(() => {
         dispatch(getAllUsers());
@@ -195,7 +17,7 @@ function UserManagement() {
 
     return(
         <>
-            <UsersHeader/>
+            <UsersHeader />
             <UsersTable users={users} isLoading={isLoading} />
         </>
     )
